refactor(adapters): support fractional broadcast timeouts and harden mongo adapter
Change BroadcastFlags.Timeout from *int64 to *float64 to preserve fractional
milliseconds from Node.js adapters across JSON, MessagePack, and BSON transports.
Update NormalizeTimerMilliseconds to accept float64, handle NaN/Infinity, and
truncate valid values to whole milliseconds within JavaScript timer bounds.

Harden MongoDB adapter by serializing publish operations through dedicated queues
to preserve insertion order and prevent reentrant deadlocks during error handling.
Improve session restoration with namespace isolation, stricter offset validation,
and correct filtering of non-recoverable packets. Fix change stream resumption by
clearing conflicting start options when applying a cached resume token. Export
PrepareClusterData to materialize binary readers before wire encoding while
preserving original payloads for local delivery.

Update documentation and tests to reflect the floating-point timeout API and
add integration coverage for Mongo publish ordering, recovery, and stream resume.

Refs: #22, #103, #136, #141

+++ adapters/redis/adapter/redis-adapter_test.go
@@ -884,7 +884,7 @@ func TestRedisAdapterTypedNilOptionsAndParserUseDefaults(t *testing.T) {
 func TestClassicBroadcastAckTimeoutNormalization(t *testing.T) {
 	tests := []struct {
 		name        string
-		timeout     int64
+		timeout     float64
 		stillActive time.Duration
 	}{
 		{name: "zero"},
